move MentionDialog to :core:commonui:content
